syscalls: Fix compiler warnings for comparison of different signedness
Fixes 45 instances of warning type:
- warning: comparison of integer expressions of different signedness

Closes: https://github.com/linux-test-project/ltp/pull/772

+++ testcases/kernel/syscalls/chown/chown05.c
@@ -128,9 +128,9 @@ int main(int ac, char **av)
 			}
 			if (stat(TESTFILE, &stat_buf) == -1)
 				tst_brkm(TFAIL, cleanup, "stat failed");
-			if (user_id == -1)
+			if ((int)user_id == -1)
 				user_id = test_cases[i - 1].user_id;
-			if (group_id == -1)
+			if ((int)group_id == -1)
 				group_id = test_cases[i - 1].group_id;
 
 			if (stat_buf.st_uid != user_id ||
