+++ testcases/kernel/syscalls/symlink/symlink01.c
@@ -598,7 +598,7 @@ struct tcses *get_tcs_info(char *ptr)
 	}
 #endif
 
-	for (ctr = 0; ctr < (sizeof(all_tcses) / sizeof(struct tcses)); ctr++) {
+	for (ctr = 0; ctr < (int)(sizeof(all_tcses) / sizeof(struct tcses)); ctr++) {
 		if (strcmp(ptr, all_tcses[ctr].tcid) == 0 ||
 		    strcmp(ptr, all_tcses[ctr].syscall) == 0) {
 			tcs_ptr = &all_tcses[ctr];
@@ -1875,7 +1875,7 @@ void cleanup(void)
 
 void help(void)
 {
-	int ind;
+	unsigned int ind;
 
 	printf("   -T id  Determines which tests cases to execute:\n");
